# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/items/attention": {
      "get": {
        "operationId": "ItemsController_attention",
        "parameters": [
          {
            "name": "withinDays",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttentionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Expiring, out-of-warranty, overdue-to-return and low-stock items everywhere",
        "tags": [
          "items"
        ]
      }
    },
    "/api/places/{placeId}/items": {
      "post": {
        "operationId": "ItemsController_create",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a thing and say which storage it went into",
        "tags": [
          "items"
        ]
      },
      "get": {
        "operationId": "ItemsController_findAll",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free text over name, aliases, tags and description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storageId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeNested",
            "required": false,
            "in": "query",
            "description": "Include items in nested storages under storageId",
            "schema": {
              "default": true,
              "type": "boolean"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "AVAILABLE",
                "LENT_OUT",
                "CONSUMED",
                "LOST",
                "DISCARDED"
              ]
            }
          },
          {
            "name": "visibility",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SHARED",
                "PRIVATE"
              ]
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expiringInDays",
            "required": false,
            "in": "query",
            "description": "Only items expiring within this many days",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lowStock",
            "required": false,
            "in": "query",
            "description": "Only items at or below their low-stock threshold",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "updatedAt",
              "type": "string",
              "enum": [
                "name",
                "createdAt",
                "updatedAt",
                "expiresAt"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedItemsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List and filter the things in a place",
        "tags": [
          "items"
        ]
      }
    },
    "/api/places/{placeId}/items/{itemId}": {
      "get": {
        "operationId": "ItemsController_findOne",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One thing, with its photos and recent movements",
        "tags": [
          "items"
        ]
      },
      "patch": {
        "operationId": "ItemsController_update",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a thing",
        "tags": [
          "items"
        ]
      },
      "delete": {
        "operationId": "ItemsController_remove",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft-delete a thing",
        "tags": [
          "items"
        ]
      }
    },
    "/api/places/{placeId}/items/{itemId}/history": {
      "get": {
        "operationId": "ItemsController_history",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ItemMovementResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Everywhere this thing has been kept",
        "tags": [
          "items"
        ]
      }
    },
    "/api/places/{placeId}/items/{itemId}/move": {
      "post": {
        "operationId": "ItemsController_move",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemDetailResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Move a thing to another storage",
        "tags": [
          "items"
        ]
      }
    },
    "/api/places/{placeId}/items/{itemId}/lend": {
      "post": {
        "operationId": "ItemsController_lend",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LendItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a thing as lent out",
        "tags": [
          "items"
        ]
      }
    },
    "/api/places/{placeId}/items/{itemId}/return": {
      "post": {
        "operationId": "ItemsController_returnItem",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a lent thing as returned",
        "tags": [
          "items"
        ]
      }
    },
...
  },
  "info": {
    "title": "FindMyStuff API",
    "description": "Remember where you kept your stuff.\n\nHierarchy: **Place** (home/office/locker) → **Storage** tree (room → almirah → shelf → box) → **Item**.\n\nEvery place-scoped route is nested under `/places/{placeId}` and is checked against your\nmembership of that place. Asking for a place you are not a member of returns **404**, not 403 —\nthe API never confirms that someone else's data exists.\n\nAll routes except signup, login, refresh, logout and health require\n`Authorization: Bearer <accessToken>`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Home — Demo deleted"
          }
        },
        "required": [
          "success"
        ]
      },
...
      "UserCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Shristi"
          },
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "avatarMediaId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
      "AttentionItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Paracetamol strip"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "expiresAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "storage": {
            "type": "object",
            "nullable": true
          },
          "place": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "quantity",
          "expiresAt"
        ]
      },
      "AttentionResponse": {
        "type": "object",
        "properties": {
          "expiring": {
            "description": "Expiring soon or already expired",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttentionItemResponse"
            }
          },
          "warranty": {
            "description": "Warranty running out",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttentionItemResponse"
            }
          },
          "overdue": {
            "description": "Lent out and past the due date",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttentionItemResponse"
            }
          },
          "lowStock": {
            "description": "At or below the low-stock threshold",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AttentionItemResponse"
            }
          }
        },
        "required": [
          "expiring",
          "warranty",
          "overdue",
          "lowStock"
        ]
      },
      "CreateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Passport"
          },
          "storageId": {
            "type": "string",
            "description": "Where it is kept. Omit if not put away yet."
          },
          "description": {
            "type": "string"
          },
          "aliases": {
            "description": "Other names you might search by, e.g. [\"charger\"] for a Type-C cable",
            "example": [
              "charger",
              "cable"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "example": [
              "documents",
              "important"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "example": "Documents"
          },
          "quantity": {
            "type": "number",
            "default": 1
          },
          "lowStockAt": {
            "type": "number",
            "description": "Warn when quantity drops to this number"
          },
          "serialNumber": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "SHARED",
              "PRIVATE"
            ],
            "default": "SHARED",
            "description": "PRIVATE hides this item from every other member of the place"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "purchasedAt": {
            "type": "string",
            "example": "2026-01-15T00:00:00.000Z"
          },
          "expiresAt": {
            "type": "string",
            "description": "Medicines, food, documents"
          },
          "warrantyUntil": {
            "type": "string"
          },
          "remindAt": {
            "type": "string",
            "description": "Remind me about this item at this time"
          },
          "imagesBase64": {
            "description": "Base64 photos of where it is kept. First one becomes the cover.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "ItemStorageRef": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Top shelf"
          },
          "breadcrumb": {
            "type": "string",
            "example": "Bedroom › Almirah › Top shelf",
            "description": "Where the thing is — render this directly"
          }
        },
        "required": [
          "id",
          "name",
          "breadcrumb"
        ]
      },
      "ItemMovementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fromStorageId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "toStorageId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "fromLabel": {
            "type": "object",
            "nullable": true,
            "example": "Bedroom › Almirah › Top shelf"
          },
          "toLabel": {
            "type": "object",
            "nullable": true,
            "example": "Bedroom › Under the bed › Blue box"
          },
          "note": {
            "type": "object",
            "nullable": true,
            "example": "Shifted while cleaning"
          },
          "movedBy": {
            "$ref": "#/components/schemas/UserCard"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fromStorageId",
          "toStorageId",
          "fromLabel",
          "toLabel",
          "note",
          "movedBy",
          "createdAt"
        ]
      },
      "ItemDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "placeId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Passport"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "example": "Inside the brown document folder"
          },
          "aliases": {
            "example": [
              "charger",
              "cable"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "example": [
              "documents",
              "important"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "object",
            "nullable": true,
            "example": "Documents"
          },
          "quantity": {
            "type": "number",
            "example": 1
          },
          "lowStockAt": {
            "type": "object",
            "nullable": true,
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "LENT_OUT",
              "CONSUMED",
              "LOST",
              "DISCARDED"
            ]
          },
          "visibility": {
            "type": "string",
            "enum": [
              "SHARED",
              "PRIVATE"
            ],
            "description": "PRIVATE is visible only to ownerId"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "warrantyUntil": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "lentToName": {
            "type": "object",
            "nullable": true,
            "example": "Rahul (neighbour)"
          },
          "dueAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "storage": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ItemStorageRef"
              }
            ]
          },
          "mediaIds": {
            "description": "Fetch each at /media/{id}/thumbnail or /media/{id}/raw",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "owner": {
            "$ref": "#/components/schemas/UserCard"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserCard"
          },
          "movements": {
            "description": "20 most recent moves",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemMovementResponse"
            }
          }
        },
        "required": [
          "id",
          "placeId",
          "name",
          "description",
          "aliases",
          "tags",
          "category",
          "quantity",
          "lowStockAt",
          "status",
          "visibility",
          "ownerId",
          "expiresAt",
          "warrantyUntil",
          "lentToName",
          "dueAt",
          "storage",
          "mediaIds",
          "updatedAt",
          "owner",
          "createdBy",
          "movements"
        ]
      },
...
      "PaginatedItemsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Passport"
          },
          "description": {
            "type": "string"
          },
          "aliases": {
            "description": "Other names you might search by, e.g. [\"charger\"] for a Type-C cable",
            "example": [
              "charger",
              "cable"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "example": [
              "documents",
              "important"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "category": {
            "type": "string",
            "example": "Documents"
          },
          "quantity": {
            "type": "number",
            "default": 1
          },
          "lowStockAt": {
            "type": "number",
            "description": "Warn when quantity drops to this number"
          },
          "serialNumber": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "SHARED",
              "PRIVATE"
            ],
            "default": "SHARED",
            "description": "PRIVATE hides this item from every other member of the place"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "purchasedAt": {
            "type": "string",
            "example": "2026-01-15T00:00:00.000Z"
          },
          "expiresAt": {
            "type": "string",
            "description": "Medicines, food, documents"
          },
          "warrantyUntil": {
            "type": "string"
          },
          "remindAt": {
            "type": "string",
            "description": "Remind me about this item at this time"
          },
          "imagesBase64": {
            "description": "Base64 photos of where it is kept. First one becomes the cover.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storageId": {
            "type": "object",
            "description": "Move the item. Pass null to mark it as not put away anywhere.",
            "nullable": true
          }
        }
      },
      "MoveItemDto": {
        "type": "object",
        "properties": {
          "toStorageId": {
            "type": "object",
            "nullable": true,
            "description": "Destination storage, or null to unassign"
          },
          "note": {
            "type": "string",
            "example": "Shifted while cleaning"
          }
        },
        "required": [
          "toStorageId"
        ]
      },
      "LendItemDto": {
        "type": "object",
        "properties": {
          "lentToName": {
            "type": "string",
            "example": "Rahul (neighbour)"
          },
          "lentToId": {
            "type": "string",
            "description": "App user id, when you lent it to someone on FindMyStuff"
          },
          "dueAt": {
            "type": "string",
            "description": "When you expect it back"
          }
        },
        "required": [
          "lentToName"
        ]
      },
...
    }
  }
}